style toolbar for draftjs text editor had to be recoded, switching from classes to function based in draftjs.js
