+++ client/package.json
@@ -5,6 +5,7 @@
   "dependencies": {
     "@draft-js-plugins/editor": "^4.1.0",
     "@draft-js-plugins/emoji": "^4.5.4",
+    "@draft-js-plugins/static-toolbar": "^4.1.1",
     "@testing-library/react": "^11.2.6",
     "@testing-library/user-event": "^12.8.3",
     "draft-js": "^0.11.7",
@@ -13,7 +14,11 @@
     "react-dom": "^17.0.2",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
-    "web-vitals": "^1.1.2"
+    "web-vitals": "^1.1.2",
+    "@giphy/js-fetch-api": "2.0.1",
+    "@giphy/react-components": "2.0.1",
+    "emoji-picker-react": "3.2.4",
+    "usetheform": "3.4.0"
   },
   "proxy": "http://localhost:3001",
   "scripts": {
